add patch avatar and resizing avatar middleware
