+++ middlewares/upload.js
@@ -1,0 +1,20 @@
+const multer = require("multer");
+const path = require("path");
+
+const storage = multer.diskStorage({
+  destination: (req, file, cb) => {
+    cb(null, path.resolve(__dirname, "../tmp"));
+  },
+  filename: (req, file, cb) => {
+    cb(null, file.originalname);
+  },
+  limits: {
+    fileSize: 2048,
+  },
+});
+
+const upload = multer({
+  storage,
+});
+
+module.exports = upload;
