+++ middlewares/validateBody.js
@@ -1,0 +1,15 @@
+const { HttpError } = require("../helpers");
+
+const validateBody = (schema) => {
+  return (req, res, next) => {
+    const { error } = schema.validate(req.body);
+
+    if (error) {
+      throw new HttpError(400, error.message);
+    }
+
+    return next();
+  };
+};
+
+module.exports = validateBody;
